+++ tests/testthat/test-balance_ions.R
@@ -89,9 +89,9 @@ test_that("Balance ions only updates TDS/cond/IS when appropriate.", {
 ################################################################################*
 # balance_ions helpers ----
 
-# Test that balance_ions_chain outputs are the same as base function, balance_ions.
+# Test that balance_ions_df outputs are the same as base function, balance_ions.
 
-test_that("balance_ions_chain outputs are the same as base function, balance_ions", {
+test_that("balance_ions_df outputs are the same as base function, balance_ions", {
   testthat::skip_on_cran()
   water1 <- suppressWarnings(define_water(
     ph = 7.9, temp = 20, alk = 50, tot_hard = 50, ca = 13, mg = 4, na = 20, k = 20,
@@ -99,8 +99,8 @@ test_that("balance_ions_chain outputs are the same as base function, balance_ion
   ))
   water2 <- balance_ions(water1)
 
-  water3 <- suppressWarnings(define_water_chain(slice(water_df, 1))) %>%
-    balance_ions_chain()
+  water3 <- suppressWarnings(define_water_df(slice(water_df, 1))) %>%
+    balance_ions_df()
 
   water4 <- purrr::pluck(water3, 2, 1)
 
@@ -112,8 +112,8 @@ test_that("balance_ions_chain outputs are the same as base function, balance_ion
   ))
   water6 <- balance_ions(water5, anion = "so4", cation = "mg")
 
-  water7 <- suppressWarnings(define_water_chain(slice(water_df, 1))) %>%
-    balance_ions_chain(anion = "so4", cation = "mg")
+  water7 <- suppressWarnings(define_water_df(slice(water_df, 1))) %>%
+    balance_ions_df(anion = "so4", cation = "mg")
 
   water8 <- purrr::pluck(water7, 2, 1)
 
@@ -122,21 +122,21 @@ test_that("balance_ions_chain outputs are the same as base function, balance_ion
 
 # Test that output is a column of water class lists, and changing the output column name works
 
-test_that("balance_ions_chain output is a column of water class lists", {
+test_that("balance_ions_df output is a column of water class lists", {
   testthat::skip_on_cran()
-  water1 <- suppressWarnings(define_water_chain(slice(water_df, 1))) %>%
-    balance_ions_chain()
+  water1 <- suppressWarnings(define_water_df(slice(water_df, 1))) %>%
+    balance_ions_df()
   water2 <- purrr::pluck(water1, 2, 1)
 
   expect_s4_class(water2, "water") # check class
 })
 
 # Check that this function can be piped to the next one
-test_that("balance_ions_chain can be piped and handle an output_water argument", {
+test_that("balance_ions_df can be piped and handle an output_water argument", {
   testthat::skip_on_cran()
-  water1 <- suppressWarnings(define_water_chain(slice(water_df, 1))) %>%
-    balance_ions_chain(output_water = "different_column") %>%
-    chemdose_ph_chain(naoh = 20)
+  water1 <- suppressWarnings(define_water_df(slice(water_df, 1))) %>%
+    balance_ions_df(output_water = "different_column") %>%
+    chemdose_ph_df(naoh = 20)
 
   expect_equal(names(water1[2]), "different_column") # check output_water arg
   expect_equal(ncol(water1), 4) # check if pipe worked